Organizing exported type declarations for use in other npm modules
