--- converters/convert.ts
@@ -1,5 +1,0 @@
-import fromarray from "./fromarray";
-import { SchemaManager } from "./schema";
-import toarray from "./toarray";
-
-export default { fromarray, toarray, SchemaManager };
