+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "clarity-js",
-  "version": "0.1.44",
+  "version": "0.2.0",
   "description": "An analytics library that uses web page interactions to generate aggregated insights",
   "author": "Microsoft Corp.",
   "license": "MIT",
-  "main": "build/src/module.js",
+  "main": "build/index.js",
+  "types": "index.d.ts",
   "keywords": [
     "clarity",
     "Microsoft",
