--- src/module.ts
@@ -1,4 +1,0 @@
-import converter from "../converters/convert";
-import * as library from "./clarity";
-
-export default { library, converter };
