+++ src/plugins/layout.ts
@@ -1,6 +1,6 @@
 import { Action, IElementLayoutState, IEventData, ILayoutRoutineInfo, ILayoutState, IMutationRoutineInfo,
   INodeInfo, Instrumentation, IPlugin, IShadowDomInconsistentEventState, IShadowDomMutationSummary, IShadowDomNode,
-  LayoutRoutine, NumberJson, Source } from "../../clarity";
+  LayoutRoutine, NumberJson, Source } from "../../types/index";
 import { config } from "./../config";
 import { addEvent, addMultipleEvents, bind, getTimestamp, instrument } from "./../core";
 import { debug, mask, traverseNodeTree } from "./../utils";
