+++ src/plugins/layout/shadowdom.ts
@@ -1,4 +1,4 @@
-import { INodeInfo, IShadowDomMutationSummary, IShadowDomNode, NumberJson } from "../../../clarity";
+import { INodeInfo, IShadowDomMutationSummary, IShadowDomNode, NumberJson } from "../../../types/index";
 import { assert, isNumber, traverseNodeTree } from "../../utils";
 import { createNodeInfo } from "./nodeinfo";
 import { getNodeIndex, NodeIndex } from "./stateprovider";
